IT WORKS
Okay I finally managed to get the estimated distance and the rssi working. Bleak using standard variables of rssi doesn't work. But I used a different value of it called advertisement_data. This works for windows and gives a fairly accurate value of how far a device is. Still does not work for apple devices but will most likely work for android.