Update to log
Whatever I typed up in my formal Google document, I put here. It is essential to say my prototype is working well, with a little bit more work needed to be done for it to meet the requirements that were set before me. I like how it works, but I know that this is still just a prototype. The usage of a DB and FIDO2 is needed for it to work properly.